Revisit and centralize shell escape and dispatching
All handlers end using the same functions to escape and dispatch to
shell.

Vim special `:!` codes, like `%` for current file, are always escaped
(`:!` itself  is the responsible to unescape them at the end).

When substituting `{{URL}}`, `{sub}` is prepended with `\=` to be
evaluated as an expression (we make it a string) and avoid codes with
special meaning. We need then to escape `"` and also `\`, or it would be
'unescaped' by `substitute()`

References #6